Move XWalkExternalExtension to common/
It will also be used by the upcoming extension process, so move to a
common area.
