+++ extensions/common/xwalk_extension_external.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "xwalk/extensions/browser/xwalk_extension_external.h"
+#include "xwalk/extensions/common/xwalk_extension_external.h"
 
 #include "base/files/file_path.h"
 #include "base/logging.h"
