+++ extensions/test/external_extension_browsertest.cc
@@ -5,8 +5,8 @@
 #include "base/native_library.h"
 #include "base/path_service.h"
 #include "base/utf_string_conversions.h"
-#include "xwalk/extensions/browser/xwalk_extension_external.h"
 #include "xwalk/extensions/browser/xwalk_extension_service.h"
+#include "xwalk/extensions/common/xwalk_extension_external.h"
 #include "xwalk/extensions/test/xwalk_extensions_test_base.h"
 #include "xwalk/runtime/browser/runtime.h"
 #include "xwalk/test/base/xwalk_test_utils.h"
