+++ runtime/browser/xwalk_browser_main_parts.cc
@@ -12,8 +12,8 @@
 #include "base/files/file_path.h"
 #include "base/strings/string_number_conversions.h"
 #include "xwalk/experimental/dialog/dialog_extension.h"
-#include "xwalk/extensions/browser/xwalk_extension_external.h"
 #include "xwalk/extensions/browser/xwalk_extension_service.h"
+#include "xwalk/extensions/common/xwalk_extension_external.h"
 #include "xwalk/runtime/browser/devtools/remote_debugging_server.h"
 #include "xwalk/runtime/browser/runtime.h"
 #include "xwalk/runtime/browser/runtime_context.h"
